==> 3_players_dictionary.py <==


d = {
     'Alonso':
         {'role' : 1,
          'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Marcos_Alonso_122/gw.csv',
          'valid_gws' : [i for i in range(1, 39)]},
    
     'Antonio':
         {'role' : 3,
          'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Michail_Antonio_413/gw.csv',
          'valid_gws' : [i for i in range(1, 39)]},
         
    'Bamford':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Patrick_Bamford_189/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    
        
    'Bernardo':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Bernardo%20Mota_Veiga%20de%20Carvalho%20e%20Silva_261/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
        
    'Bowen':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Jarrod_Bowen_420/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
        
    'Calvert-Lewin':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Dominic_Calvert-Lewin_177/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    
    'Cancelo':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Jo%C3%A3o%20Pedro%20Cavaco_Cancelo_256/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
     
    'Coutinho':
       {'role' : 2 ,
       'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Philippe_Coutinho%20Correia_681/gw.csv',
       'valid_gws' : [i for i in range(1, 39)]},   
        
    'De Breuyne':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Kevin_De%20Bruyne_251/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]}, 
        
    'Dennis':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Emmanuel_Dennis_450/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
         
     'Salah':
    {'role' : 2,
     'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Mohamed_Salah_233/gw.csv',
     'valid_gws' : [i for i in range(1, 39)]},
    
    'Diaz':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Luis_D%C3%ADaz_696/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Digne':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Lucas_Digne_168/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Fabianski':
        {'role' : 0,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Lukasz_Fabianski_408/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Fernandes':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Bruno%20Miguel_Borges%20Fernandes_277/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Foden':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Phil_Foden_265/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Gallagher':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Conor_Gallagher_144/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Havertz':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Kai_Havertz_141/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Iheanacho':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Kelechi_Iheanacho_213/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Ings':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Danny_Ings_337/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'James':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Reece_James_142/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Jesus':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Gabriel%20Fernando_de%20Jesus_263/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Jimenez':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Ra%C3%BAl_Jim%C3%A9nez_430/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Jota':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Diogo_Jota_240/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Kane':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Harry_Kane_357/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Lacazette':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Alexandre_Lacazette_6/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Livramento':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Tino_Livramento_491/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Lukaku':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Romelu_Lukaku_529/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Mahrez':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Riyad_Mahrez_254/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Mané':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Sadio_Man%C3%A9_230/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Martinez':
        {'role' : 0,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Emiliano_Mart%C3%ADnez_30/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Maupay':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Neal_Maupay_63/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Mount':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Mason_Mount_138/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Pukki':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Teemu_Pukki_315/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Ramsdale':
        {'role' : 0,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Aaron_Ramsdale_559/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Raphinha':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Raphael_Dias%20Belloli_196/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Rashford':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Marcus_Rashford_281/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Reguilon':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Sergio_Reguil%C3%B3n_370/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Robertson':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Andrew_Robertson_234/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Ronaldo':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Cristiano%20Ronaldo_dos%20Santos%20Aveiro_579/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Saka':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Bukayo_Saka_22/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Shaw':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Luke_Shaw_275/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Son':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Heung-Min_Son_359/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Sterling':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Raheem_Sterling_255/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'TAA':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Trent_Alexander-Arnold_237/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Toney':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Ivan_Toney_78/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Van Dijk':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Virgil_van%20Dijk_229/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Vardy':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Jamie_Vardy_205/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Watkins':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Ollie_Watkins_40/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Zaha':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Wilfried_Zaha_154/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Martial':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Anthony_Martial_278/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Trossard':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Leandro_Trossard_64/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Gündogan':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Ilkay_G%C3%BCndogan_250/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Benrahma':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Sa%C3%AFd_Benrahma_419/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Werner':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Timo_Werner_134/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Chilwell':
        {'role' : 1,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Benjamin_Chilwell_135/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Cavani':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Edinson_Cavani_269/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'H. Barnes':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Harvey_Barnes_215/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Sancho':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Jadon_Sancho_485/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Greenwood':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Mason_Greenwood_289/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Aubameyang':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Pierre-Emerick_Aubameyang_4/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Wilson':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Callum_Wilson_296/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Grealish':
        {'role' : 2,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Jack_Grealish_33/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]},
    'Firmino':
        {'role' : 3,
         'url' : 'https://raw.githubusercontent.com/vaastav/Fantasy-Premier-League/master/data/2021-22/players/Roberto_Firmino_228/gw.csv',
         'valid_gws' : [i for i in range(1, 39)]}
    
    }
